chore: move checkbox atom out of watch menu
